Reorganize type definition
